+++ app/forms/transfer/task/deed_of_variation_task_form.rb
@@ -1,0 +1,8 @@
+class Transfer::Task::DeedOfVariationTaskForm < BaseOptionalTaskForm
+  attribute :received, :boolean
+  attribute :cleared, :boolean
+  attribute :signed, :boolean
+  attribute :saved, :boolean
+  attribute :sent, :boolean
+  attribute :signed_secretary_state, :boolean
+end